Integrated 'add switch statements' into the new mutation framework.
